ipq40xx: Support for 512MB Giga spi nand
(cherry picked from commit 19b016d80ea231fa6552c4d784cfd7f77a9e7220)

+++ drivers/staging/mt29f_spinand/mt29f_spinand.h
@@ -101,10 +101,10 @@ struct spinand_ops {
 	void (*spinand_set_defaults)(struct spi_device *spi_nand);
 	void (*spinand_read_cmd)(struct spinand_cmd *cmd, u32 page_id);
 	void (*spinand_read_data)(struct spinand_cmd *cmd, u16 column,
-				  u16 page_id);
+				  u32 page_id);
 	void (*spinand_write_cmd)(struct spinand_cmd *cmd, u32 page_id);
 	void (*spinand_write_data)(struct spinand_cmd *cmd, u16 column,
-				   u16 page_id);
+				   u32 page_id);
 	void (*spinand_erase_blk)(struct spinand_cmd *cmd, u32 page_id);
 	int (*spinand_parse_id)(struct spi_device *spi_nand, u8 *nand_id,
 				u8 *id);
